Web_search_Agent/requirements.txt
python-dotenv  
tavily-python
pydantic
Ipython
fastapi
uvicorn